requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.5
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-mathfilters==1.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-starfield==1.0.post2
filelock==3.0.12
gunicorn==20.1.0
mysql-connector-python==8.0.26
Pillow==8.3.2
pipenv==2020.11.15
psycopg2==2.9.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
virtualenv==20.3.1
virtualenv-clone==0.5.4
whitenoise==5.2.0
